+++ cpp/src/io/reader.cpp
@@ -80,20 +80,20 @@ CopcInfo Reader::ReadCopcInfoVlr()
 CopcExtents Reader::ReadCopcExtentsVlr(std::map<uint64_t, las::VlrHeader> &vlrs, const las::EbVlr &eb_vlr) const
 {
     auto offset = FetchVlr(vlrs, "copc", 10000);
+    auto extended_offset = FetchVlr(vlrs, "rock_robotic", 10001);
     if (offset == 0)
         throw std::runtime_error("Reader::ReadCopcExtentsVlr: No COPC Extents VLR found in file.");
 
     in_stream_->seekg(offset + lazperf::vlr_header::Size);
     CopcExtents extents(las::CopcExtentsVlr::create(*in_stream_, vlrs[offset].data_length),
                         static_cast<int8_t>(reader_->header().point_format_id),
-                        static_cast<uint16_t>(eb_vlr.items.size()));
+                        static_cast<uint16_t>(eb_vlr.items.size()), extended_offset != 0);
 
     // Load mean/var if they exist
-    offset = FetchVlr(vlrs, "rock_robotic", 10001);
-    if (offset != 0)
+    if (extended_offset != 0)
     {
-        in_stream_->seekg(offset + lazperf::vlr_header::Size);
-        extents.UpdateExtendedStats(las::CopcExtentsVlr::create(*in_stream_, vlrs[offset].data_length));
+        in_stream_->seekg(extended_offset + lazperf::vlr_header::Size);
+        extents.UpdateExtendedStats(las::CopcExtentsVlr::create(*in_stream_, vlrs[extended_offset].data_length));
     }
     return extents;
 }