+++ cpp/src/io/writer_internal.cpp
@@ -18,24 +18,27 @@ size_t WriterInternal::OffsetToPointData() const
 {
     size_t base_offset(375 + (54 + 160) + (54 + (34 + 4 * 6))); // header + COPC vlr + LAZ vlr (max 4 items)
 
-    size_t extents_offset = CopcExtents::ByteSize(copc_file_writer_->LasHeader()->PointFormatId(),
-                                                  copc_file_writer_->ExtraBytesVlr().items.size()) +
-                            lazperf::vlr_header::Size;
+    size_t extents_offset =
+        CopcExtents::ByteSize(copc_config_->LasHeader()->PointFormatId(), copc_config_->ExtraBytesVlr().items.size()) +
+        lazperf::vlr_header::Size;
+    // If we store extended stats we need two extents VLRs
+    if (copc_config_->CopcExtents()->HasExtendedStats())
+        extents_offset *= 2;
 
-    size_t wkt_offset = copc_file_writer_->Wkt().size();
+    size_t wkt_offset = copc_config_->Wkt().size();
     if (wkt_offset > 0)
         wkt_offset += lazperf::vlr_header::Size;
 
-    size_t eb_offset = copc_file_writer_->ExtraBytesVlr().size();
+    size_t eb_offset = copc_config_->ExtraBytesVlr().size();
     if (eb_offset > 0)
         eb_offset += lazperf::vlr_header::Size;
 
-    return base_offset + 2 * extents_offset + wkt_offset + eb_offset; // 2*extents to store extended stats
+    return base_offset + extents_offset + wkt_offset + eb_offset;
 }
 
 WriterInternal::WriterInternal(std::ostream &out_stream, std::shared_ptr<CopcConfigWriter> copc_config,
                                std::shared_ptr<Hierarchy> hierarchy)
-    : out_stream_(out_stream), copc_file_writer_(copc_config), hierarchy_(hierarchy)
+    : out_stream_(out_stream), copc_config_(copc_config), hierarchy_(hierarchy)
 {
     // reserve enough space for the header & VLRs in the file
     std::fill_n(std::ostream_iterator<char>(out_stream_), FirstChunkOffset(), 0);
@@ -66,55 +69,58 @@ void WriterInternal::Close()
 // Writes the LAS header and VLRs
 void WriterInternal::WriteHeader()
 {
-    laz_vlr lazVlr(copc_file_writer_->LasHeader()->PointFormatId(), copc_file_writer_->LasHeader()->EbByteSize(),
+    laz_vlr lazVlr(copc_config_->LasHeader()->PointFormatId(), copc_config_->LasHeader()->EbByteSize(),
                    VARIABLE_CHUNK_SIZE);
 
-    auto las_header_vlr = copc_file_writer_->LasHeader()->ToLazPerf(OffsetToPointData(), point_count_, evlr_offset_,
-                                                                    evlr_count_, !copc_file_writer_->Wkt().empty(),
-                                                                    copc_file_writer_->LasHeader()->EbByteSize());
+    auto las_header_vlr = copc_config_->LasHeader()->ToLazPerf(
+        OffsetToPointData(), point_count_, evlr_offset_, evlr_count_, !copc_config_->Wkt().empty(),
+        copc_config_->LasHeader()->EbByteSize(), copc_config_->CopcExtents()->HasExtendedStats());
 
     out_stream_.seekp(0);
     las_header_vlr.write(out_stream_);
 
     // Write the COPC Info VLR.
-    auto copc_info_vlr = copc_file_writer_->CopcInfo()->ToLazPerf();
+    auto copc_info_vlr = copc_config_->CopcInfo()->ToLazPerf();
     copc_info_vlr.header().write(out_stream_);
     copc_info_vlr.write(out_stream_);
 
     // Write the COPC Extents VLR.
-    auto extents_vlr = copc_file_writer_->CopcExtents()->ToLazPerf({las_header_vlr.minx, las_header_vlr.maxx},
-                                                                   {las_header_vlr.miny, las_header_vlr.maxy},
-                                                                   {las_header_vlr.minz, las_header_vlr.maxz});
+    auto extents_vlr = copc_config_->CopcExtents()->ToLazPerf({las_header_vlr.minx, las_header_vlr.maxx},
+                                                              {las_header_vlr.miny, las_header_vlr.maxy},
+                                                              {las_header_vlr.minz, las_header_vlr.maxz});
     extents_vlr.header().write(out_stream_);
     extents_vlr.write(out_stream_);
 
     // Write the COPC Extended stats VLR.
-    auto extended_stats_vlr = copc_file_writer_->CopcExtents()->ToLazPerfExtended();
+    if (copc_config_->CopcExtents()->HasExtendedStats())
+    {
+        auto extended_stats_vlr = copc_config_->CopcExtents()->ToLazPerfExtended();
 
-    auto header = extended_stats_vlr.header();
-    header.user_id = "rock_robotic";
-    header.record_id = 10001;
-    header.description = "COPC extended stats";
+        auto header = extended_stats_vlr.header();
+        header.user_id = "rock_robotic";
+        header.record_id = 10001;
+        header.description = "COPC extended stats";
 
-    header.write(out_stream_);
-    extended_stats_vlr.write(out_stream_);
+        header.write(out_stream_);
+        extended_stats_vlr.write(out_stream_);
+    }
 
     // Write the LAZ VLR
     lazVlr.header().write(out_stream_);
     lazVlr.write(out_stream_);
 
     // Write optional WKT VLR
-    if (!copc_file_writer_->Wkt().empty())
+    if (!copc_config_->Wkt().empty())
     {
-        lazperf::wkt_vlr wkt_vlr(copc_file_writer_->Wkt());
+        lazperf::wkt_vlr wkt_vlr(copc_config_->Wkt());
         wkt_vlr.header().write(out_stream_);
         wkt_vlr.write(out_stream_);
     }
 
     // Write optional Extra Byte VLR
-    if (copc_file_writer_->LasHeader()->EbByteSize() > 0)
+    if (copc_config_->LasHeader()->EbByteSize() > 0)
     {
-        auto ebVlr = this->copc_file_writer_->ExtraBytesVlr();
+        auto ebVlr = this->copc_config_->ExtraBytesVlr();
         ebVlr.header().write(out_stream_);
         ebVlr.write(out_stream_);
     }
@@ -170,7 +176,7 @@ Entry WriterInternal::WriteNode(std::vector<char> in, int32_t point_count, bool 
     if (compressed)
         out_stream_.write(in.data(), in.size());
     else
-        point_count = laz::Compressor::CompressBytes(out_stream_, *copc_file_writer_->LasHeader(), in);
+        point_count = laz::Compressor::CompressBytes(out_stream_, *copc_config_->LasHeader(), in);
 
     point_count_ += point_count;
 
@@ -210,8 +216,8 @@ void WriterInternal::WritePage(const std::shared_ptr<PageInternal> &page)
     // Set the copc header info if needed
     if (page->key == VoxelKey::BaseKey())
     {
-        copc_file_writer_->CopcInfo()->root_hier_offset = offset;
-        copc_file_writer_->CopcInfo()->root_hier_size = page_size;
+        copc_config_->CopcInfo()->root_hier_offset = offset;
+        copc_config_->CopcInfo()->root_hier_size = page_size;
     }
 
     for (const auto &node : page->nodes)