+++ test/reader_node_test.cpp
@@ -88,7 +88,7 @@ TEST_CASE("GetPoints Test", "[Reader] ")
         REQUIRE(points[0]->ScanAngle() == 1667);
         REQUIRE_THAT(points[0]->ScanAngleDegrees(), Catch::Matchers::WithinAbs(10, 0.01));
         REQUIRE(points[0]->UserData() == 124);
-        REQUIRE(points[0]->PointSourceID() == 7327);
+        REQUIRE(points[0]->PointSourceId() == 7327);
         REQUIRE_THAT(points[0]->GPSTime(), Catch::Matchers::WithinAbs(246098.8141472744, 0.0001));
         REQUIRE(points[0]->Red() == 16896);
         REQUIRE(points[0]->Green() == 65280);
@@ -110,7 +110,7 @@ TEST_CASE("GetPoints Test", "[Reader] ")
         points[0]->Classification(31);
         points[0]->ScanAngle(90);
         points[0]->UserData(std::numeric_limits<int8_t>::max());
-        points[0]->PointSourceID(std::numeric_limits<int8_t>::max());
+        points[0]->PointSourceId(std::numeric_limits<int8_t>::max());
         points[0]->GPSTime(std::numeric_limits<double>::max());
         points[0]->Red(std::numeric_limits<uint16_t>::max());
         points[0]->Green(std::numeric_limits<uint16_t>::max());
@@ -128,7 +128,7 @@ TEST_CASE("GetPoints Test", "[Reader] ")
         REQUIRE(points[0]->Classification() == 31);
         REQUIRE(points[0]->ScanAngle() == 90);
         REQUIRE(points[0]->UserData() == std::numeric_limits<int8_t>::max());
-        REQUIRE(points[0]->PointSourceID() == std::numeric_limits<int8_t>::max());
+        REQUIRE(points[0]->PointSourceId() == std::numeric_limits<int8_t>::max());
         REQUIRE(points[0]->GPSTime() == std::numeric_limits<double>::max());
         REQUIRE(points[0]->Red() == std::numeric_limits<uint16_t>::max());
         REQUIRE(points[0]->Green() == std::numeric_limits<uint16_t>::max());